Allowing short circuit selector evaluation
There was an issue with short circuiting stateful operations from
ranges.  The previous version of romulus just eagerly evaluated both
sides of the selector expression.  Now the stateful statement is
performed when the line is processed.  This enables selector operation
nodes to skip evaluation of the secondary node and the interpreter still
understands the state even though the total expression may block that
line from being processed